add some icon
add some icon and debug xjx pull request code.
