+++ app/src/main/res/menu/main.xml
@@ -4,19 +4,19 @@
     <item
         android:id="@+id/action_search"
         android:actionViewClass="android.widget.SearchView"
-        android:icon="@drawable/action_search"
+        android:icon="@drawable/action_search_l"
         app:showAsAction="ifRoom|collapseActionView"
         android:title="@string/action_search"/>
     <item
         android:id="@+id/action_add"
         android:orderInCategory="100"
-        android:icon="@drawable/ic_menu_send"
+        android:icon="@drawable/social_send_now"
         android:title="@string/action_settings"
         app:showAsAction="never" />
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
-        android:icon="@drawable/ic_menu_share"
+        android:icon="@drawable/social_share"
         android:title="@string/action_settings"
         app:showAsAction="never" />
 </menu>
